pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coco2hf-format"
version = "0.1.0"
description = "Convert COCO format annotations to HuggingFace dataset metadata format"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "huggingface-hub>=1.2.3",
    "pillow>=12.1.0",
    "ruff>=0.14.10",
    "tqdm>=4.67.1",
]

[project.scripts]
coco2hf = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
]
